captioncraft/backend/requirements.txt
# CaptionCraft Backend Dependencies

# Web Framework
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-multipart==0.0.9

# AWS S3 SDK (for Cloudflare R2)
boto3==1.34.34
botocore==1.34.34

# AssemblyAI SDK
assemblyai==0.23.1

# HTTP Client
httpx==0.26.0
aiohttp==3.9.3

# Video/Audio Processing (FFmpeg wrapper)
ffmpeg-python==0.2.0

# Utilities
python-dotenv==1.0.1
pydantic==2.6.1
pydantic-settings==2.1.0

# Async support
aiofiles==23.2.1

# Optional: Job queue (uncomment if scaling)
# redis==5.0.1
# rq==1.16.0
